Fix atlascode Dockerfile. Make build.sh better (add clean option, argument checking... Add description to update.sh

+++ build/build.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/bin/bash -e
 #
 # Copyright (c) 2023 Red Hat, Inc.
 # This program and the accompanying materials are made
@@ -7,8 +7,30 @@
 #
 # SPDX-License-Identifier: EPL-2.0
 #
+# Contributors:
+#   Red Hat, Inc. - initial API and implementation
+#
 
-set -e
+CLEAN=0
+
+usage ()
+{
+    echo "Usage: $0 EXTENSION_NAME [--clean]
+
+Example: $0 atlassian.atlascode --clean
+
+Options: 
+  --clean     : Run podman system prune -a -f after the build to clean up leftover containers
+"
+    exit
+}
+
+if [[ -z "$1" ]]; then usage; fi
+if [[ ! -z "$2" ]] && [[ $2 = "--clean" ]]; then
+    CLEAN=1
+else
+    usage
+fi
 
 EXTENSION_NAME="$1"
 
@@ -71,8 +93,11 @@ BUILDER_CONTAINER_ID=$(cat "$EXTENSION_NAME"-builder-id)
 podman cp $BUILDER_CONTAINER_ID:/$EXTENSION_NAME.vsix ./
 podman cp $BUILDER_CONTAINER_ID:/$EXTENSION_NAME-sources.tar.gz ./
 podman stop $BUILDER_CONTAINER_ID
-podman system prune -a -f
 rm ./$EXTENSION_NAME-builder-id
+
+if [[ $CLEAN -eq 1 ]]; then
+    podman system prune -a -f
+fi
 
 # Get SHA256 of vsix and sources files and add to plugin-manifest.json for the Brew build
 PLUGIN_SHA=$(sha256sum $EXTENSION_NAME.vsix)
